Work on unit tests
