+++ src/uwtools/tests/config/test_core.py
@@ -1,6 +1,6 @@
-# pylint: disable=duplicate-code,missing-function-docstring,protected-access,redefined-outer-name
+# pylint: disable=missing-function-docstring,protected-access,redefined-outer-name
 """
-Tests for uwtools.config module.
+Tests for uwtools.config.core module.
 """
 
 import datetime
@@ -56,54 +56,6 @@ def test_compare_config(caplog, fmt, salad_base):
         "salad:            size:  - None + large",
     ]:
         assert logged(caplog, msg)
-
-
-def test_compare_configs_good(compare_configs_assets, caplog):
-    log.setLevel(logging.INFO)
-    _, a, b = compare_configs_assets
-    assert tools.compare_configs(
-        config_a_path=a, config_a_format=FORMAT.yaml, config_b_path=b, config_b_format=FORMAT.yaml
-    )
-    assert caplog.records
-
-
-def test_compare_configs_changed_value(compare_configs_assets, caplog):
-    log.setLevel(logging.INFO)
-    d, a, b = compare_configs_assets
-    d["baz"]["qux"] = 11
-    with writable(b) as f:
-        yaml.dump(d, f)
-    assert not tools.compare_configs(
-        config_a_path=a, config_a_format=FORMAT.yaml, config_b_path=b, config_b_format=FORMAT.yaml
-    )
-    assert logged(caplog, "baz:             qux:  - 99 + 11")
-
-
-def test_compare_configs_missing_key(compare_configs_assets, caplog):
-    log.setLevel(logging.INFO)
-    d, a, b = compare_configs_assets
-    del d["baz"]
-    with writable(b) as f:
-        yaml.dump(d, f)
-    # Note that a and b are swapped:
-    assert not tools.compare_configs(
-        config_a_path=b, config_a_format=FORMAT.yaml, config_b_path=a, config_b_format=FORMAT.yaml
-    )
-    assert logged(caplog, "baz:             qux:  - None + 99")
-
-
-def test_compare_configs_bad_format(caplog):
-    log.setLevel(logging.INFO)
-    with raises(UWConfigError) as e:
-        tools.compare_configs(
-            config_a_path="/not/used",
-            config_a_format="jpg",
-            config_b_path="/not/used",
-            config_b_format=FORMAT.yaml,
-        )
-    msg = "Format 'jpg' should be one of: fieldtable, ini, nml, yaml"
-    assert logged(caplog, msg)
-    assert msg in str(e.value)
 
 
 def test_config_field_table(tmp_path):
@@ -815,18 +767,6 @@ def test_YAMLConfig__load_paths_failure_stdin_plus_relpath(caplog):
 
 
 # Helper functions
-
-
-@fixture
-def compare_configs_assets(tmp_path):
-    d = {"foo": {"bar": 88}, "baz": {"qux": 99}}
-    a = tmp_path / "a"
-    b = tmp_path / "b"
-    with writable(a) as f:
-        yaml.dump(d, f)
-    with writable(b) as f:
-        yaml.dump(d, f)
-    return d, a, b
 
 
 def help_realize_config_fmt2fmt(infn, infmt, cfgfn, cfgfmt, tmpdir):
